Rename 'OfferingViewing' screen to 'Home'
