+++ src/screens/Analytics/index.js
@@ -1,21 +1,28 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { CTLayout } from 'components';
+import { api } from 'utils';
 
-export function Analytics() {
-  const layoutProps = CTLayout.createProps({
-    transition: true,
-    responsive: true,
-    footer: true,
-    headingProps: {
-      heading: 'Personal Analytics',
-      icon: 'bar_chart',
-      sticky: true,
-      gradient: true,
-      offsetTop: 30
-    }
-  });
+export class Analytics extends Component {
+  componentDidMount() {
+    api.contentLoaded();
+  }
 
-  return (
-    <CTLayout {...layoutProps} />
-  )
+  render() {
+    const layoutProps = CTLayout.createProps({
+      transition: true,
+      responsive: true,
+      footer: true,
+      headingProps: {
+        heading: 'Personal Analytics',
+        icon: 'bar_chart',
+        sticky: true,
+        gradient: true,
+        offsetTop: 30
+      }
+    });
+
+    return (
+      <CTLayout {...layoutProps} />
+    )
+  }
 }
