+++ src/screens/Home/Components/Home/Section/OfferingCard.jsx
