+++ src/screens/Home/Components/Home/Section/StarredSection.jsx
