+++ src/screens/Home/Components/OfferingDetail/PlaylistsView.jsx
