+++ src/screens/Home/Components/OfferingDetail/VideoView.jsx
