+++ src/screens/Home/Components/OfferingDetail/index.css
