+++ src/screens/Home/Components/OfferingDetail/index.js
