+++ src/screens/Home/Components/index.js
