+++ src/screens/Home/Utils/user-guide.js
